data-loader-genie/requirements.txt
pyyaml
nipyapi